use newer debian version in Docker images
Unfortunately, it seems like the `golang:1.21.3` images and the
`gcr.io/distroless/base` images are based on different debian versions,
so in order to keep them consistent, we have to specify both debian
versions explicitly.

That means when a new stable debian is released, we'll have to update
the versions here explicitly and can't rely on Dependabot to update them
ourselves.

The `golang` images use the Debian code name (bookworm), while the
distroless images uses its version number (12).