migrate from OpenCensus to OpenTelemetry (#582)
This changes how we collect metrics and create traces. The old way was
to use OpenCensus. The new way is to use OpenTelemetry.

OTel comes with some tooling so we don't need to do some of the GCP API
calls we were doing before. It also has a bug with handling metrics when
a context is errored that we discuss and workaround in
recordErrorMetric.

We do a minimal swap. So, we leave code using metrics as using the new
otel metrics and do not try to turn metrics into trace calls.

The aggregation View configs seem to no longer be required as the
defaults for those metrics ("Instruments") are now good (at least, in
the stdouttrace and stdoutmetric cases). So, we remove them.

The new `must*` functions for the Instruments are annoying but okay.
Also, setting up the MeterProvider and TraceProvider with
`otel.SetMeterProvider` and `otel.SetTraceProvider` after using the
`otel.Meter` and `otel.Tracer` functions work as you'd like. This is any
Instruments or Views made with those global values will be copied over
to the new provider. However, the actual values (that is, actual counter
numbers) won't be, so metrics shouldn't be incremented or modified til
after the `Set*Provider` calls are made.

I'm not sure what would be a better way of doing these refactorings, but
we're trying to stay minimal so keeping the counters and gauges as
globals seems fine for now.