copy config struct in confLoader.Get (#583)
There was no race condition, because the config struct was never being
modified and we always recreated it whole cloth, but this makes sure
everyone's happy long term. We get rid of a related FIXME meanwhile.

Along the way, we get to drop all the JSON specific stuff (including
pointers for nil value support) from the config struct that most of the
code is handling.