use a distro-agnostic Go image tag
Since we're very unlikely to have Ubuntu distribution-specific issues with our Go
build, we can use the distribution-agnostic Go image tag.

This will make it possible for dependabot to update our Go image tag
even when the new Go version is only on a newer Ubuntu distribution.

We could likely use the distroless Go docker image instead, but this
minimizes the change for now. Future work!